Password Generator
Creating a strong password and remembering it is a tedious task. You can build a program that intakes some words from the user and then generates a random password using those words. The user can remember the password with the help of the words he gave as an input.
Take the Input from user.
Using that input try to generate Random password.
Save that password into text file.
Tell the User is Generated password strong or weak.